fix: Remove the 4GB stream ceiling; prove the Go/Dart window divergence
Follow-up to d9b1dc7, addressing P0, P1 and P4 of
RECOMMENDATIONS_FROM_GO_RICOCHET.md.

Serial-number arithmetic for the advertised offset
--------------------------------------------------

d9b1dc7 clamped the advertised offset at the uint32 wire ceiling. go-ricochet
asked what a stream actually does on arrival there, which was the right
question: it stalls permanently. That is the original 256KB bug with a bigger
number on it, and 4GB is reachable for a long-lived connection moving blobs.

The offset now travels modulo 2^32 and the sender reconstructs it against the
limit it already holds (RFC 1982). This is unambiguous because the true offset
is always within one receive window of the current limit, and the window is
capped at 4MB against a 4GB modulus — a margin of some 500x. No wire change,
so the ceiling is gone without any coordination with dart-udx. That removes one
of the two items P4 wanted scheduled together; only the per-stream offset in
StreamFrame still needs a wire revision.

Proving the Go/Dart divergence
------------------------------

P0 was inferred from reading and flagged as unproven. It is now proven, by
running the real Dart implementation rather than reasoning about it:

  Go consumed 243,048 bytes; 2,030,400 buffered unread; advertised window 524,288
  → the Dart sender ran 3.9x past the offset it was granted

dart-udx's sender gates on `inflight < _remoteReceiveWindow` (stream.dart:170,
:273, :344), comparing bytes currently outstanding against an offset that grows
for the life of the stream, so the comparison stops binding almost immediately
and cwnd becomes the only limit on a Dart->Go upload. `inflight` is also the
connection-level counter, so it is the wrong quantity for a per-stream limit
regardless.

The fix is smaller than the report assumed, and is written up in
doc/DART_SENDER_PATCH.md: no wire change is needed, because both
implementations already *send* absolute offsets. dart-udx's receiver advertises
65536 + cumulativeBytesReceived (stream.dart:228), which is an offset with a
constant 64KB slack — it is only named like a window size. So Go comparing its
lifetime dataSent against it is correct, not coincidental, and only Dart's
sender-side interpretation is wrong. The patch also covers the wrap, and is
derived from reading dart-udx@5291cdb without being compiled or run.

Bulk interop suite
------------------

The interop suite moved 2,000 and 4,096 bytes, never reaching a window update,
which is how a 256KB ceiling survived in a project that already had a netem
harness. Adds interop/dartpeer, a bulk Dart peer living here rather than in
dart-udx because it is go-udx test scaffolding, plus three tests: sustained
multi-megabyte transfer in both directions, and the slow-consumer case.

TestBulk_DartToGoSlowConsumer asserts the exact invariant a compliant sender
must satisfy — BufferedBytes() <= RecvWindow(), since we advertise
dataConsumed + recvWindow. It FAILS TODAY, deliberately: it is the proof of the
dart-udx defect and goes green when that patch lands. The other two pass.

Also adds the Go-to-Go form as TestTransfer_SlowConsumerAppliesBackpressure,
which passes: a reader consuming 161KB held its sender to 226KB with the
receive buffer peaking at 60KB.

Exports Stream.BufferedBytes and Stream.RecvWindow so the invariant can be
asserted from outside the package.

P2 and P5 needed no change. The grep P2 suggested turns up one other
GetPacket-guarded-by-nil at connection.go:774, but DetectLostPackets does not
delete, so that guard is legitimate. The e2e controller-state assertions P2
asked for already exist in congestion_wiring_test.go. P5's comment is at
flow_control.go:5.

P3 stands: the 8MB/2.17s and 64MB/54MB-s figures were loopback with a
congestion controller that had been live for minutes. They are evidence of
"not broken", not a baseline. Re-measuring under netem is still outstanding.

+++ connection.go
@@ -267,7 +267,11 @@ func (c *Connection) sendResetStream(streamID, remoteID uint32, errorCode uint32
 // sendWindowUpdate advertises an ABSOLUTE offset limit for the stream: the
 // highest cumulative byte position the peer may send. It is not a window size.
 // See StreamFlowController for why absolute offsets are required here.
-func (c *Connection) sendWindowUpdate(streamID, remoteID uint32, maxStreamData int) {
+//
+// The uint32 conversion is a deliberate truncation: the offset travels modulo
+// 2^32 and the peer reconstructs it against the limit it already holds, which
+// is what keeps a stream's lifetime transfer unbounded over a 4-byte field.
+func (c *Connection) sendWindowUpdate(streamID, remoteID uint32, maxStreamData int64) {
 	frame := &WindowUpdateFrame{WindowSize: uint32(maxStreamData)}
 	c.sendPacket(remoteID, streamID, []Frame{frame})
 }
@@ -697,7 +701,7 @@ func (c *Connection) handleFrame(pkt *Packet, frame Frame) {
 		// inflate our receive buffer just by claiming to be blocked.
 		if s := c.findStream(pkt.DestinationStreamID, pkt.SourceStreamID); s != nil {
 			if s.streamFC != nil {
-				c.sendWindowUpdate(s.ID, s.RemoteID, int(s.streamFC.RefreshLimit()))
+				c.sendWindowUpdate(s.ID, s.RemoteID, s.streamFC.RefreshLimit())
 			}
 		}
 	}
@@ -780,7 +784,7 @@ func (c *Connection) handleAckFrame(f *AckFrame) {
 func (c *Connection) handleWindowUpdate(pkt *Packet, f *WindowUpdateFrame) {
 	s := c.findStream(pkt.DestinationStreamID, pkt.SourceStreamID)
 	if s != nil {
-		s.OnWindowUpdate(int64(f.WindowSize))
+		s.OnWindowUpdate(f.WindowSize)
 	}
 }
